fix: 각 자동차의 이름에 대한 검증 기능을 Car 객체에서 Name 객체로 이동
