+++ src/main/java/racingcar/domain/car/Car.java
@@ -3,73 +3,50 @@ package racingcar.domain.car;
 import racingcar.domain.GoOrStop;
 
 public class Car implements Comparable<Car> {
-    private static final int MAX_NAME_LENGTH = 5;
     private static final String IS = " : ";
 
     private final Name name;
-    private Location location;
+    private Position position;
 
     public Car(String name) {
-        validateEmptyName(name);
-        validateNameLength(name);
-
         this.name = new Name(name);
-        this.location = new Location();
+        this.position = new Position();
     }
 
     public Car(String name, int location) {
-        validateEmptyName(name);
-        validateNameLength(name);
-
         this.name = new Name(name);
-        this.location = new Location(location);
-    }
-
-    public static void validateEmptyName(String input) throws IllegalArgumentException {
-        if (input.isEmpty()) {
-            throw new IllegalArgumentException("입력값은 빈 문자열일 수 없습니다.");
-        }
-    }
-
-    public static void validateNameLength(String input) {
-        if (isOverLength(input)) {
-            throw new IllegalArgumentException("자동차의 이름은 5자이하만 가능합니다.");
-        }
-    }
-
-    private static boolean isOverLength(String input) {
-        return input.length() > MAX_NAME_LENGTH;
+        this.position = new Position(location);
     }
 
     public void play() {
-        if (GoOrStop.decideGoOrStop()) {
+        if (GoOrStop.decideToGo()) {
             move();
         }
     }
 
     private void move() {
-        location.moveForward();
+        position.moveForward();
     }
 
     public String getRoundResult() {
         StringBuilder result = new StringBuilder();
         result.append(name);
         result.append(IS);
-        result.append(location);
+        result.append(position);
         return result.toString();
     }
 
     public boolean isAtSameLocation(Car anotherCar) {
-        return this.location.compareTo(anotherCar.location) == 0;
+        return this.position.compareTo(anotherCar.position) == 0;
     }
 
     @Override
     public int compareTo(Car anotherCar) {
-        if (this.location.compareTo(anotherCar.location) > 0) {
+        if (this.position.compareTo(anotherCar.position) > 0) {
             return 1;
         }
 
-        if (this.location.compareTo(anotherCar.location) < 0) {
+        if (this.position.compareTo(anotherCar.position) < 0) {
             return -1;
         }
